requirements.txt
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
notion-client==2.2.1
yagmail==0.15.293
beautifulsoup4==4.12.2
pandas==2.1.4
google-generativeai==0.3.1
canvasapi==3.2.0 